requirements.txt
mcp>=1.0.0
libclang>=16.0.0